+++ app/src/app/Shared/Services/Translator/Utils/conversions.ts
@@ -1,3 +1,5 @@
+// src/app/Shared/Utils/conversions.ts
+
 /**
  * Convierte una cadena binaria a su representación hexadecimal.
  * @param binaryString La cadena binaria a convertir.
