+++ app/src/app/Shared/Services/Translator/test/aritmetic-Itype-inst-test.spec.ts
@@ -1,6 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { TranslatorService } from '../app/Shared/Services/Translator/translator.service';
-
+import { TranslatorService } from '../translator.service';
 
 describe('TranslatorService aritmetic I-type instruction tests', () => {
     let service: TranslatorService;
