+++ app/tsconfig.spec.json
@@ -10,6 +10,6 @@
   },
   "include": [
     "src/**/*.spec.ts",
-    "src/**/*.d.ts",
+    "src/**/*.d.ts"
   ]
 }
